+++ tailwind.config.ts
@@ -9,8 +9,11 @@ export default {
   theme: {
     extend: {
       colors: {
-        background: "var(--background)",
-        foreground: "var(--foreground)",
+        black: '#000000',
+        blue: '#14213D',
+        yellow: '#FCA311',
+        grey: '#E5E5E5',
+        white: '#FFFFFF',
       },
     },
   },
